pyproject.toml
[project]
name = "tsc-inspection-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.1",
    "google-genai>=1.49.0",
    "uvicorn>=0.38.0",
]
